requirements.txt
pandas>=1.5
pyyaml
tenacity
pytest
numpy
